work on binning op ui

+++ beam-binning2/src/main/java/org/esa/beam/binning/operator/ui/VariableConfigTable.java
@@ -141,7 +141,7 @@ class VariableConfigTable {
         return scrollPane;
     }
 
-    void addRow(final String name, String expression, String algorithmName, double weightCoefficient, double fillValue) {
+    void addRow(final String name, String expression, String algorithmName, double weightCoefficient, float fillValue) {
         if (algorithmName == null || !StringUtils.contains(aggregatorNames, algorithmName)) {
             algorithmName = AggregatorAverage.Descriptor.NAME;
         }
